Update Authentifation file

+++ src/controllers/specialiteController.js
@@ -126,36 +126,56 @@ const specialiteController = {
     }
   },
 
-  // Fonction pour supprimer une spécialité
-  deleteSpecialite: async (req, res) => {
-    const { id } = req.params
+// Fonction pour supprimer une spécialité
+deleteSpecialite: async (req, res) => {
+  const { id } = req.params
 
-    try {
-      const specialite = await prisma.specialites.findUnique({
-        where: { id: Number(id) }
-      })
+  try {
+    // Vérification si la spécialité existe
+    const specialite = await prisma.specialites.findUnique({
+      where: { id: Number(id) }
+    })
 
-      if (!specialite) {
-        return res.status(404).json({ message: 'Spécialité non trouvée' })
-      }
-
-      await prisma.specialites.delete({
-        where: { id: Number(id) }
-      })
-
-      return res
-        .status(200)
-        .json({ message: 'Spécialité supprimée avec succès' })
-    } catch (error) {
-      console.error(error)
-      return res
-        .status(500)
-        .json({
-          message: 'Erreur lors de la suppression de la spécialité',
-          error
-        })
+    if (!specialite) {
+      return res.status(404).json({ message: 'Spécialité non trouvée' })
     }
+
+    // Vérification si la spécialité est associée à un utilisateur
+    const utilisateurAssocie = await prisma.utilisateurs.findFirst({
+      where: {
+        specialite_id: Number(id)  // Utilisation du bon nom de champ (specialite_id)
+      }
+    })
+
+    if (utilisateurAssocie) {
+      return res.status(400).json({ message: 'Impossible de supprimer la spécialité car elle est associée à un utilisateur.' })
+    }
+
+    // Suppression de la spécialité
+    await prisma.specialites.delete({
+      where: { id: Number(id) }
+    })
+
+    return res
+      .status(200)
+      .json({ message: 'Spécialité supprimée avec succès' })
+  } catch (error) {
+    console.error(error)
+    if (error instanceof Prisma.PrismaClientValidationError) {
+      return res.status(400).json({
+        message: 'Erreur de validation lors de la suppression',
+        error: error.message
+      })
+    }
+    return res
+      .status(500)
+      .json({
+        message: 'Erreur lors de la suppression de la spécialité',
+        error
+      })
   }
+}
+
 }
 
 export default specialiteController
